Refactor: created style folder for css files
