+++ src/screens/spell_book.jsx
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router-dom";
 import SpellTable from "../components/spell_table";
-import "./spell_book.css"
+import "../styles/spell_book.css"
 
 
 export default function SpellBook() {
