+++ src/styles/spell_card.css
