+++ src/styles/spell_table.css
